requirements.txt
bs4==0.0.2
colorlog==6.9.0
crfm-helm==0.5.5
docker==7.1.0
fastapi==0.115.6
openai==1.70
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.8.0
uvicorn[standard]==0.34.0
pre-commit==4.1.0
# NOTE: The versions of black and isort are hardcoded here.
# If you change them, you must also update .pre-commit-config.yaml and .github/workflows/ci.yml.
black==24.10.0
flake8==7.1.1
isort==5.13.2
coverage==7.6.10
pytest==7.4.4
pytest-asyncio==0.23.6
pytest-mock==3.12.0
mkdocs==1.6.1
selenium==4.16.0
webdriver-manager==4.0.2
anthropic>=0.5.0
google-generativeai>=0.3.0
together>=0.1.5
flask-wtf>=1.0.0
